refactor(contracts): decompose ContractsService into Query/Lifecycle/Cancellation services
Behavior-preserving money-path. Facade keeps 12-method surface + 5-arg ctor +
delegates; internally constructs 3 sub-services (specs+module untouched, private
cancellationTemplate hack preserved). 5 $transaction moved WHOLE (create
retry-loop+error-switch as one unit; approveCancellation reversal-JE $tx atomic);
money math compute+persist paired. tsc 0, contracts tests green, lint 0.
